+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2022-09-03
+
+### Added
+
+- `Encrypter` class that encrypts ids into 128-bit values using AES encryption
+  of Web Crypto API.
+
+### Changed
+
+- Renamed `ObjectId64` class into `Encoder`.
+
 ## [3.1.1] - 2022-08-03
 
 ### Changed